refactored out part usage, normalized operation and schema lib structure
