+++ gql/lib/src/schema/definitions.dart
@@ -1,4 +1,7 @@
-export "./definitions/definitions.dart"
+export "./definitions/base_types.dart";
+export "./definitions/definitions.dart";
+export "./definitions/type_resolver.dart";
+export "./definitions/value_types.dart"
     hide
         // while it is defined in value_types.dart so that Value.fromNode can be made complete,
         // Variables are not part of the schema definition language
