+++ gql/lib/src/schema/definitions/type_resolver.dart
@@ -1,4 +1,7 @@
-part of "definitions.dart";
+import "package:meta/meta.dart";
+import "./base_types.dart" show TypeDefinition;
+import "./definitions.dart"
+    show InterfaceTypeDefinition, ObjectTypeDefinition, UnionTypeDefinition;
 
 /// Callback to dereference a full type definition by name
 typedef ResolveType = TypeDefinition Function(String name);
